Implement Express backend with session management and dynamic routing
