+++ Readme.MD
@@ -41,3 +41,26 @@ You can run the site with a backend so content is stored in PostgreSQL and edita
    `npm run seed`  
    Default admin: username `admin`, password `admin123` (or set `ADMIN_PASSWORD` in `.env`).
 4. **Run**: `npm install` then `npm start`. Open http://localhost:3000 for the site and http://localhost:3000/admin to log in and edit content.
+
+## Deploy to Netlify (backend + dynamic “index”)
+
+The repo is set up so the **entire backend runs on Netlify** as a serverless function. The site’s home page (“index”) is **not** a static file; it’s rendered from the database on every request, so updates in the admin panel are reflected as soon as someone visits the site.
+
+### Netlify setup
+
+1. **Connect the repo** to Netlify (GitHub/GitLab/Bitbucket).
+2. **Build settings** (optional): build command can stay empty or use `echo 'Functions only'`; publish directory = `public`.
+3. **Environment variables** (Site settings → Environment variables):
+   - `DATABASE_URL` – your PostgreSQL connection string (e.g. from Neon, Supabase, Railway).
+   - `SESSION_SECRET` – a long random string for admin sessions.
+   - `ADMIN_PASSWORD` – (optional) used when you run `npm run seed`; set before seeding the production DB.
+4. **Deploy**. All traffic is redirected to the Express app (see `netlify.toml`). The home page and `/admin` are served by the backend; content comes from the database.
+
+### One-time database setup for production
+
+From your machine (with `DATABASE_URL` in `.env` pointing at the **production** DB):
+
+- `npm run init-db` – create tables  
+- `npm run seed` – insert default content and admin user  
+
+After that, the Netlify site will read the same DB; changes in the admin will show up on the next page load.
